spilt pkg/eden.go into hypervisor specific files
