+++ pkg/eden/eden.go
@@ -1,7 +1,6 @@
 package eden
 
 import (
-	"bufio"
 	"bytes"
 	"crypto/tls"
 	"encoding/json"
@@ -360,420 +359,6 @@ func StatusEServer() (status string, err error) {
 	return state, nil
 }
 
-//StartEVEVBox function run EVE in VirtualBox
-func StartEVEVBox(vmName, eveImageFile string, cpus int, mem int, hostFwd map[string]string, ipMap map[string]net.IP) (err error) {
-	poweroff := false
-	if out, _, err := utils.RunCommandAndWait("VBoxManage", strings.Fields(fmt.Sprintf("showvminfo %s --machinereadable", vmName))...); err != nil {
-		log.Info("No VMs with eve_live name", err)
-		commandArgsString := fmt.Sprintf("createvm --name %s --register", vmName)
-		if err = utils.RunCommandWithLogAndWait("VBoxManage", defaults.DefaultLogLevelToPrint, strings.Fields(commandArgsString)...); err != nil {
-			log.Fatalf("VBoxManage error for command %s %s", commandArgsString, err)
-		}
-		commandArgsString = fmt.Sprintf("modifyvm %s --cpus %d --memory %d --vram 16 --nested-hw-virt on --ostype Ubuntu_64  --mouse usbtablet --graphicscontroller vmsvga --boot1 disk --boot2 net", vmName, cpus, mem)
-		if err = utils.RunCommandWithLogAndWait("VBoxManage", defaults.DefaultLogLevelToPrint, strings.Fields(commandArgsString)...); err != nil {
-			log.Fatalf("VBoxManage error for command %s %s", commandArgsString, err)
-		}
-
-		commandArgsString = fmt.Sprintf("storagectl %s --name \"SATA\" --add sata --bootable on --hostiocache on", vmName)
-		if err = utils.RunCommandWithLogAndWait("VBoxManage", defaults.DefaultLogLevelToPrint, strings.Fields(commandArgsString)...); err != nil {
-			log.Fatalf("VBoxManage error for command %s %s", commandArgsString, err)
-		}
-
-		commandArgsString = fmt.Sprintf("storageattach %s  --storagectl \"SATA\" --port 0 --device 0 --type hdd --medium %s", vmName, eveImageFile)
-		if err = utils.RunCommandWithLogAndWait("VBoxManage", defaults.DefaultLogLevelToPrint, strings.Fields(commandArgsString)...); err != nil {
-			log.Fatalf("VBoxManage error for command %s %s", commandArgsString, err)
-		}
-
-		commandArgsString = fmt.Sprintf("natnetwork add --netname %s --network %s --enable --dhcp on",
-										"natnet1", defaults.DefaultVBoxSubnet)
-		if err = utils.RunCommandWithLogAndWait("VBoxManage", defaults.DefaultLogLevelToPrint, strings.Fields(commandArgsString)...); err != nil {
-			log.Fatalf("VBoxManage error for command %s %s", commandArgsString, err)
-		}
-		commandArgsString = fmt.Sprintf("natnetwork start --netname %s", "natnet1")
-		if err = utils.RunCommandWithLogAndWait("VBoxManage", defaults.DefaultLogLevelToPrint, strings.Fields(commandArgsString)...); err != nil {
-			log.Errorf("VBoxManage error for command %s %s", commandArgsString, err)
-		}
-
-		commandArgsString = fmt.Sprintf("modifyvm %s --nic1 natnetwork --nat-network1 %s --cableconnected1 on",
-										vmName, "natnet1")
-		if err = utils.RunCommandWithLogAndWait("VBoxManage", defaults.DefaultLogLevelToPrint, strings.Fields(commandArgsString)...); err != nil {
-			log.Fatalf("VBoxManage error for command %s %s", commandArgsString, err)
-		}
-
-		eth0IP := ipMap["eth0"]
-		for k, v := range hostFwd {
-			commandArgsString = fmt.Sprintf("natnetwork  modify --netname %s --port-forward-4 :tcp:[]:%s:[%s]:%s",
-											"natnet1", k, eth0IP.String(), v)
-			if err = utils.RunCommandWithLogAndWait("VBoxManage", defaults.DefaultLogLevelToPrint, strings.Fields(commandArgsString)...); err != nil {
-				log.Fatalf("VBoxManage error for command %s %s", commandArgsString, err)
-			}
-		}
-
-		commandArgsString = fmt.Sprintf("modifyvm %s  --nic2 natnetwork --nat-network2 %s --cableconnected2 on",
-										vmName, "natnet1")
-		if err = utils.RunCommandWithLogAndWait("VBoxManage", defaults.DefaultLogLevelToPrint, strings.Fields(commandArgsString)...); err != nil {
-			log.Fatalf("VBoxManage error for command %s %s", commandArgsString, err)
-		}
-		eth1IP := ipMap["eth1"]
-		for k, v := range hostFwd {
-			hostPort, err := strconv.Atoi(k)
-			if err != nil {
-				log.Errorf("Parsing %s to Integer value failed", k)
-				break
-			}
-			hostPort += 10
-			guestPort, err := strconv.Atoi(v)
-			if err != nil {
-				log.Errorf("Parsing %s to Integer value failed", v)
-				break
-			}
-			guestPort += 10
-			commandArgsString = fmt.Sprintf("natnetwork  modify --netname %s --port-forward-4 :tcp:[]:%d:[%s]:%d",
-											"natnet1", hostPort, eth1IP.String(), guestPort)
-			if err = utils.RunCommandWithLogAndWait("VBoxManage", defaults.DefaultLogLevelToPrint, strings.Fields(commandArgsString)...); err != nil {
-				log.Fatalf("VBoxManage error for command %s %s", commandArgsString, err)
-			}
-		}
-
-		commandArgsString = fmt.Sprintf("startvm  %s", vmName)
-		if err = utils.RunCommandWithLogAndWait("VBoxManage", defaults.DefaultLogLevelToPrint, strings.Fields(commandArgsString)...); err != nil {
-			log.Fatalf("VBoxManage error for command %s %s", commandArgsString, err)
-		}
-	} else {
-		scanner := bufio.NewScanner(bytes.NewReader([]byte(out)))
-		for scanner.Scan() {
-			line := strings.TrimSpace(scanner.Text())
-			if strings.Contains(line, "VMState=\"poweroff\"") {
-				poweroff = true
-				break
-			}
-		}
-
-		networkFound := false
-		output := ""
-		var err error
-		if output, _, err = utils.RunCommandAndWait("VBoxManage", strings.Fields(fmt.Sprintf("natnetwork list  %s", "natnet1"))...); err != nil {
-			log.Fatalf("VBoxManage error for command natnetwork list %s", err)
-		}
-		scanner = bufio.NewScanner(bytes.NewReader([]byte(output)))
-		for scanner.Scan() {
-			line := strings.TrimSpace(scanner.Text())
-			if strings.HasPrefix(line, "Enabled") {
-				enabled := strings.Split(line, ":")[1]
-				enabled = strings.TrimSpace(enabled)
-				if enabled == "Yes" {
-					networkFound = true
-				}
-			}
-		}
-		if !networkFound {
-			commandArgsString := fmt.Sprintf("natnetwork add --netname %s --network %s --enable --dhcp on",
-			"natnet1", "10.0.2.0/24")
-			if err = utils.RunCommandWithLogAndWait("VBoxManage", defaults.DefaultLogLevelToPrint, strings.Fields(commandArgsString)...); err != nil {
-				log.Fatalf("VBoxManage error for command %s %s", commandArgsString, err)
-			}
-			commandArgsString = fmt.Sprintf("natnetwork start --netname %s", "natnet1")
-			if err = utils.RunCommandWithLogAndWait("VBoxManage", defaults.DefaultLogLevelToPrint, strings.Fields(commandArgsString)...); err != nil {
-				log.Errorf("VBoxManage error for command %s %s", commandArgsString, err)
-			}
-
-			commandArgsString = fmt.Sprintf("modifyvm %s --nic1 natnetwork --nat-network1 %s --cableconnected1 on",
-			vmName, "natnet1")
-			if err = utils.RunCommandWithLogAndWait("VBoxManage", defaults.DefaultLogLevelToPrint, strings.Fields(commandArgsString)...); err != nil {
-				log.Fatalf("VBoxManage error for command %s %s", commandArgsString, err)
-			}
-
-			commandArgsString = fmt.Sprintf("modifyvm %s --nic2 natnetwork --nat-network2 %s --cableconnected2 on",
-			vmName, "natnet1")
-			if err = utils.RunCommandWithLogAndWait("VBoxManage", defaults.DefaultLogLevelToPrint, strings.Fields(commandArgsString)...); err != nil {
-				log.Fatalf("VBoxManage error for command %s %s", commandArgsString, err)
-			}
-		}
-
-		eth0IP := ipMap["eth0"]
-		for k, v := range hostFwd {
-			commandArgsString := fmt.Sprintf("natnetwork  modify --netname %s --port-forward-4 :tcp:[]:%s:[%s]:%s",
-											"natnet1", k, eth0IP.String(), v)
-			if err = utils.RunCommandWithLogAndWait("VBoxManage", defaults.DefaultLogLevelToPrint, strings.Fields(commandArgsString)...); err != nil {
-				if err.Error() == "exit status 2" {
-					log.Infof("VBoxManage NAT rule: %s: already exists", commandArgsString)
-				} else {
-					log.Fatalf("VBoxManage error for command %s %s", commandArgsString, err)
-				}
-			}
-		}
-
-		eth1IP := ipMap["eth1"]
-		for k, v := range hostFwd {
-			hostPort, err := strconv.Atoi(k)
-			if err != nil {
-				log.Errorf("Parsing %s to Integer value failed", k)
-				break
-			}
-			hostPort += 10
-			guestPort, err := strconv.Atoi(v)
-			if err != nil {
-				log.Errorf("Parsing %s to Integer value failed", v)
-				break
-			}
-			guestPort += 10
-			commandArgsString := fmt.Sprintf("natnetwork  modify --netname %s --port-forward-4 :tcp:[]:%d:[%s]:%d",
-											"natnet1", hostPort, eth1IP.String(), guestPort)
-
-			if err = utils.RunCommandWithLogAndWait("VBoxManage", defaults.DefaultLogLevelToPrint, strings.Fields(commandArgsString)...); err != nil {
-				if err.Error() == "exit status 2" {
-					log.Infof("VBoxManage NAT rule: %s: already exists", commandArgsString)
-				} else {
-					log.Fatalf("VBoxManage error for command %s %s", commandArgsString, err)
-				}
-			}
-		}
-		if poweroff {
-			commandArgsString := fmt.Sprintf("startvm  %s", vmName)
-			if err = utils.RunCommandWithLogAndWait("VBoxManage", defaults.DefaultLogLevelToPrint, strings.Fields(commandArgsString)...); err != nil {
-				log.Fatalf("VBoxManage error for command %s %s", commandArgsString, err)
-			}
-		}
-	}
-
-	return err
-}
-
-//StartEVEQemu function run EVE in qemu
-func StartEVEQemu(qemuARCH, qemuOS, eveImageFile, qemuSMBIOSSerial string, eveTelnetPort int, qemuHostFwd map[string]string, qemuAccel bool, qemuConfigFile, logFile string, pidFile string, foregroud bool) (err error) {
-	qemuCommand := ""
-	qemuOptions := "-display none -nodefaults -no-user-config "
-	qemuOptions += fmt.Sprintf("-serial chardev:char0 -chardev socket,id=char0,port=%d,host=localhost,server,nodelay,nowait,telnet,logfile=%s ", eveTelnetPort, logFile)
-	if qemuSMBIOSSerial != "" {
-		qemuOptions += fmt.Sprintf("-smbios type=1,serial=%s ", qemuSMBIOSSerial)
-	}
-	nets, err := utils.GetSubnetsNotUsed(1)
-	if err != nil {
-		return fmt.Errorf("StartEVEQemu: %s", err)
-	}
-	offset := 0
-	network := nets[0].Subnet
-	qemuOptions += fmt.Sprintf("-netdev user,id=eth%d,net=%s,dhcpstart=%s", 0, network, nets[0].FirstAddress)
-	for k, v := range qemuHostFwd {
-		origPort, err := strconv.Atoi(k)
-		if err != nil {
-			log.Errorf("Failed converting %s to Integer", k)
-			break
-		}
-		newPort, err := strconv.Atoi(v)
-		if err != nil {
-			log.Errorf("Failed converting %s to Integer", v)
-			break
-		}
-		qemuOptions += fmt.Sprintf(",hostfwd=tcp::%d-:%d", origPort + offset, newPort + offset)
-	}
-	qemuOptions += fmt.Sprintf(" -device virtio-net-pci,netdev=eth%d ", 0)
-	offset += 10
-
-	qemuOptions += fmt.Sprintf("-netdev user,id=eth%d,net=%s,dhcpstart=%s", 1, network, nets[0].SecondAddress)
-	for k, v := range qemuHostFwd {
-		origPort, err := strconv.Atoi(k)
-		if err != nil {
-			log.Errorf("Failed converting %s to Integer", k)
-			break
-		}
-		newPort, err := strconv.Atoi(v)
-		if err != nil {
-			log.Errorf("Failed converting %s to Integer", v)
-			break
-		}
-		qemuOptions += fmt.Sprintf(",hostfwd=tcp::%d-:%d", origPort + offset, newPort + offset)
-	}
-	qemuOptions += fmt.Sprintf(" -device virtio-net-pci,netdev=eth%d ", 1)
-
-	if qemuOS == "" {
-		qemuOS = runtime.GOOS
-	} else {
-		qemuOS = strings.ToLower(qemuOS)
-	}
-	if qemuOS != "linux" && qemuOS != "darwin" {
-		return fmt.Errorf("StartEVEQemu: OS not supported: %s", qemuOS)
-	}
-	if qemuARCH == "" {
-		qemuARCH = runtime.GOARCH
-	} else {
-		qemuARCH = strings.ToLower(qemuARCH)
-	}
-	switch qemuARCH {
-	case "amd64":
-		qemuCommand = "qemu-system-x86_64"
-		if qemuAccel {
-			if qemuOS == "darwin" {
-				qemuOptions += defaults.DefaultQemuAccelDarwin
-			} else {
-				qemuOptions += defaults.DefaultQemuAccelLinuxAmd64
-			}
-		} else {
-			qemuOptions += "--cpu SandyBridge "
-		}
-	case "arm64":
-		qemuCommand = "qemu-system-aarch64"
-		if qemuAccel {
-			qemuOptions += defaults.DefaultQemuAccelLinuxArm64
-		}
-	default:
-		return fmt.Errorf("StartEVEQemu: Arch not supported: %s", qemuARCH)
-	}
-	qemuOptions += fmt.Sprintf("-drive file=%s,format=qcow2 ", eveImageFile)
-	if qemuConfigFile != "" {
-		qemuOptions += fmt.Sprintf("-readconfig %s ", qemuConfigFile)
-	}
-	log.Infof("Start EVE: %s %s", qemuCommand, qemuOptions)
-	if foregroud {
-		if err := utils.RunCommandForeground(qemuCommand, strings.Fields(qemuOptions)...); err != nil {
-			return fmt.Errorf("StartEVEQemu: %s", err)
-		}
-	} else {
-		log.Infof("With pid: %s ; log: %s", pidFile, logFile)
-		if err := utils.RunCommandNohup(qemuCommand, logFile, pidFile, strings.Fields(qemuOptions)...); err != nil {
-			return fmt.Errorf("StartEVEQemu: %s", err)
-		}
-	}
-	return nil
-}
-
-//StopEVEQemu function stop EVE
-func StopEVEQemu(pidFile string) (err error) {
-	return utils.StopCommandWithPid(pidFile)
-}
-
-//StopEVEVBox function stop EVE in VirtualBox
-func StopEVEVBox(vmName string) (err error) {
-	commandArgsString := fmt.Sprintf("natnetwork modify --netname %s --dhcp off", "natnet1")
-	if err = utils.RunCommandWithLogAndWait("VBoxManage", defaults.DefaultLogLevelToPrint, strings.Fields(commandArgsString)...); err != nil {
-		log.Errorf("VBoxManage error for command %s %s", commandArgsString, err)
-	}
-	commandArgsString = fmt.Sprintf("natnetwork stop --netname %s", "natnet1")
-	if err = utils.RunCommandWithLogAndWait("VBoxManage", defaults.DefaultLogLevelToPrint, strings.Fields(commandArgsString)...); err != nil {
-		log.Errorf("VBoxManage error for command %s %s", commandArgsString, err)
-	}
-	commandArgsString = fmt.Sprintf("natnetwork remove --netname %s", "natnet1")
-	if err = utils.RunCommandWithLogAndWait("VBoxManage", defaults.DefaultLogLevelToPrint, strings.Fields(commandArgsString)...); err != nil {
-		log.Errorf("VBoxManage error for command %s %s", commandArgsString, err)
-	}
-	commandArgsString = fmt.Sprintf("controlvm %s poweroff", vmName)
-	if err = utils.RunCommandWithLogAndWait("VBoxManage", defaults.DefaultLogLevelToPrint, strings.Fields(commandArgsString)...); err != nil {
-		log.Errorf("VBoxManage error for command %s %s", commandArgsString, err)
-	} else {
-		for i := 0; i < 5; i++ {
-			time.Sleep(defaults.DefaultRepeatTimeout)
-			status, err := StatusEVEVBox(vmName)
-			if err != nil {
-				return err
-			}
-			if strings.Contains(status, "poweroff") {
-				return nil
-			}
-		}
-	}
-	return err
-}
-
-//DeleteEVEVBox function removes EVE from VirtualBox
-func DeleteEVEVBox(vmName string) (err error) {
-	commandArgsString := fmt.Sprintf("unregistervm %s --delete", vmName)
-	if err = utils.RunCommandWithLogAndWait("VBoxManage", defaults.DefaultLogLevelToPrint, strings.Fields(commandArgsString)...); err != nil {
-		log.Errorf("VBoxManage error for command %s %s", commandArgsString, err)
-	}
-	return err
-}
-
-//DeleteEVEParallels function removes EVE from parallels
-func DeleteEVEParallels(vmName string) (err error) {
-	commandArgsString := fmt.Sprintf("delete %s", vmName)
-	if err = utils.RunCommandWithLogAndWait("prlctl", defaults.DefaultLogLevelToPrint, strings.Fields(commandArgsString)...); err != nil {
-		log.Errorf("prlctl error for command %s %s", commandArgsString, err)
-	}
-	return err
-}
-
-//StatusEVEQemu function get status of EVE
-func StatusEVEQemu(pidFile string) (status string, err error) {
-	return utils.StatusCommandWithPid(pidFile)
-}
-
-//StartEVEParallels function run EVE in parallels
-func StartEVEParallels(vmName, eveImageFile string, parallelsCpus int, parallelsMem int, hostFwd map[string]string) (err error) {
-	status, err := StatusEVEParallels(vmName)
-	if err != nil {
-		log.Fatal(err)
-	}
-	if strings.Contains(status, "running") {
-		return nil
-	}
-	_ = StopEVEParallels(vmName)
-
-	commandArgsString := fmt.Sprintf("create %s --distribution ubuntu --no-hdd", vmName)
-	if err = utils.RunCommandWithLogAndWait("prlctl", defaults.DefaultLogLevelToPrint, strings.Fields(commandArgsString)...); err != nil {
-		log.Fatalf("prlctl error for command %s %s", commandArgsString, err)
-	}
-	commandArgsString = fmt.Sprintf("set %s --device-del net0 --cpus %d --memsize %d --nested-virt on --adaptive-hypervisor on --hypervisor-type parallels", vmName, parallelsCpus, parallelsMem)
-	if err = utils.RunCommandWithLogAndWait("prlctl", defaults.DefaultLogLevelToPrint, strings.Fields(commandArgsString)...); err != nil {
-		log.Fatalf("prlctl error for command %s %s", commandArgsString, err)
-	}
-	dirForParallels := strings.TrimRight(eveImageFile, filepath.Ext(eveImageFile))
-	commandArgsString = fmt.Sprintf("set %s --device-add hdd --image %s", vmName, dirForParallels)
-	if err = utils.RunCommandWithLogAndWait("prlctl", defaults.DefaultLogLevelToPrint, strings.Fields(commandArgsString)...); err != nil {
-		log.Fatalf("prlctl error for command %s %s", commandArgsString, err)
-	}
-	commandArgsString = fmt.Sprintf("set %s --device-add net --type shared --adapter-type virtio --ipadd 192.168.1.0/24 --dhcp yes", vmName)
-	if err = utils.RunCommandWithLogAndWait("prlctl", defaults.DefaultLogLevelToPrint, strings.Fields(commandArgsString)...); err != nil {
-		log.Fatalf("prlctl error for command %s %s", commandArgsString, err)
-	}
-	commandArgsString = fmt.Sprintf("set %s --device-add net --type shared --adapter-type virtio --ipadd 192.168.2.0/24 --dhcp yes", vmName)
-	if err = utils.RunCommandWithLogAndWait("prlctl", defaults.DefaultLogLevelToPrint, strings.Fields(commandArgsString)...); err != nil {
-		log.Fatalf("prlctl error for command %s %s", commandArgsString, err)
-	}
-	for k, v := range hostFwd {
-		commandArgsString = fmt.Sprintf("net set Shared --nat-tcp-add %s_%s,%s,%s,%s", k, v, k, vmName, v)
-		if err = utils.RunCommandWithLogAndWait("prlsrvctl", defaults.DefaultLogLevelToPrint, strings.Fields(commandArgsString)...); err != nil {
-			log.Fatalf("prlsrvctl error for command %s %s", commandArgsString, err)
-		}
-	}
-	commandArgsString = fmt.Sprintf("start %s", vmName)
-	return utils.RunCommandWithLogAndWait("prlctl", defaults.DefaultLogLevelToPrint, strings.Fields(commandArgsString)...)
-}
-
-//StopEVEParallels function stop EVE and delete parallels VM
-func StopEVEParallels(vmName string) (err error) {
-	commandArgsString := fmt.Sprintf("stop %s --kill", vmName)
-	if err = utils.RunCommandWithLogAndWait("prlctl", defaults.DefaultLogLevelToPrint, strings.Fields(commandArgsString)...); err != nil {
-		log.Errorf("prlctl error for command %s %s", commandArgsString, err)
-	}
-	return DeleteEVEParallels(vmName)
-}
-
-//StatusEVEVBox function get status of EVE
-func StatusEVEVBox(vmName string) (status string, err error) {
-	out, _, err := utils.RunCommandAndWait("VBoxManage", strings.Fields(fmt.Sprintf("showvminfo %s --machinereadable", vmName))...)
-	if err != nil {
-		return "", err
-	}
-	scanner := bufio.NewScanner(bytes.NewReader([]byte(out)))
-	for scanner.Scan() {
-		line := strings.TrimSpace(scanner.Text())
-		if strings.Contains(line, "VMState=") {
-			return strings.Split(line, "\"")[1], nil
-		}
-	}
-	return "process doesn''t exist", nil
-}
-
-//StatusEVEParallels function get status of EVE
-func StatusEVEParallels(vmName string) (status string, err error) {
-	commandArgsString := fmt.Sprintf("status %s", vmName)
-	statusEVE, _, err := utils.RunCommandAndWait("prlctl", strings.Fields(commandArgsString)...)
-	if err != nil {
-		return "process doesn''t exist", nil
-	}
-	statusEVE = strings.TrimLeft(statusEVE, fmt.Sprintf("VM %s exist ", vmName))
-	return statusEVE, nil
-}
 
 //GenerateEveCerts function generates certs for EVE
 func GenerateEveCerts(certsDir, domain, ip, eveIP, uuid, devModel, ssid, password string) (err error) {
